refactor: Protocol Extension을 사용해서 MainTableViewDelegate, MainTableViewDataSource 에서 함수 분리
